+++ api/src/test/java/marquez/common/api/JobResourceIntegrationTest.java
@@ -12,6 +12,7 @@ import com.google.common.collect.ImmutableSet;
 import java.util.List;
 import java.util.UUID;
 import marquez.BaseIntegrationTest;
+import marquez.api.JdbiUtils;
 import marquez.client.MarquezHttpException;
 import marquez.client.models.DbTableMeta;
 import marquez.client.models.Job;
@@ -43,22 +44,7 @@ public class JobResourceIntegrationTest extends BaseIntegrationTest {
 
   @AfterEach
   public void tearDown(Jdbi jdbi) {
-    jdbi.inTransaction(
-        handle -> {
-          handle.execute("DELETE FROM lineage_events");
-          handle.execute("DELETE FROM runs_input_mapping");
-          handle.execute("DELETE FROM dataset_versions_field_mapping");
-          handle.execute("DELETE FROM stream_versions");
-          handle.execute("DELETE FROM dataset_versions");
-          handle.execute("UPDATE runs SET start_run_state_uuid=NULL, end_run_state_uuid=NULL");
-          handle.execute("DELETE FROM run_states");
-          handle.execute("DELETE FROM runs");
-          handle.execute("DELETE FROM run_args");
-          handle.execute("DELETE FROM job_versions_io_mapping");
-          handle.execute("DELETE FROM job_versions");
-          handle.execute("DELETE FROM jobs");
-          return null;
-        });
+    JdbiUtils.cleanDatabase(jdbi);
   }
 
   @Test